+++ WebApi/Contexts/Interfaces/IWebContext.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 using WebApi.Models;
@@ -12,48 +13,48 @@ namespace WebApi.Contexts.Interfaces
         bool ValidateConnection();
         #region Product
 
-        WebProduct GetProductById(int id);
+        WebProduct GetProductById(BigInteger id);
         IEnumerable<WebProduct> GetAllProducts();
-        IEnumerable<WebProduct> GetAllProductsExcludingIds(IList<int> idsToExclude);
-        int CreateProduct(WebProduct entity, List<ExcludedFields> excludedFields = null);
-        int UpdateProduct(int id, WebProduct entity, List<ExcludedFields> excludedFields = null);
-        int GetTotalProductsCount(string searchTerm);
+        IEnumerable<WebProduct> GetAllProductsExcludingIds(IList<BigInteger> idsToExclude);
+        BigInteger CreateProduct(WebProduct entity, List<ExcludedFields> excludedFields = null);
+        BigInteger UpdateProduct(BigInteger id, WebProduct entity, List<ExcludedFields> excludedFields = null);
+        BigInteger GetTotalProductsCount(string searchTerm);
         IEnumerable<WebProduct> GetAllProductsWithFields(ProductTypes productType);
         IEnumerable<WebProduct> SearchProducts(string searchTerm, ProductTypes productType, int page = 1, int pageSize = 10, int maxPage = 1);
-        int GetMaxProductId();
-        int GetTotalCategoriesCount(string searchTerm);
+        BigInteger GetMaxProductId();
+        BigInteger GetTotalCategoriesCount(string searchTerm);
 
         #endregion
 
         #region  VariableProduct
-        int CreateVariableProduct(WebProduct variableProduct, List<ExcludedFields> excludedFields = null);
-        int CreateVariationProduct(int variableId, WebProduct variationProduct, List<ExcludedFields> excludedFields = null);
+        BigInteger CreateVariableProduct(WebProduct variableProduct, List<ExcludedFields> excludedFields = null);
+        BigInteger CreateVariationProduct(BigInteger variableId, WebProduct variationProduct, List<ExcludedFields> excludedFields = null);
         IEnumerable<WebProduct> GetVariableProductsBySearch(string searchTerm);
-        IEnumerable<WebProduct> GetVariationProductsByVariableId(int variableId);
-        void UpdateVariationProduct(int variableId, WebProduct variationProduct, List<ExcludedFields> excludedFields = null);
+        IEnumerable<WebProduct> GetVariationProductsByVariableId(BigInteger variableId);
+        void UpdateVariationProduct(BigInteger variableId, WebProduct variationProduct, List<ExcludedFields> excludedFields = null);
 
         #endregion
 
         #region Category
-        WebCategory GetCategoryById(int id);
+        WebCategory GetCategoryById(BigInteger id);
         IEnumerable<WebCategory> GetAllCategories();
-        IEnumerable<WebCategory> GetAllCategoriesExcludingIds(IList<int> idsToExclude);
-        int CreateCategory(WebCategory entity, List<ExcludedFields> excludedFields = null);
-        int UpdateCategory(int id, WebCategory entity, List<ExcludedFields> excludedFields = null);
+        IEnumerable<WebCategory> GetAllCategoriesExcludingIds(IList<BigInteger> idsToExclude);
+        BigInteger CreateCategory(WebCategory entity, List<ExcludedFields> excludedFields = null);
+        BigInteger UpdateCategory(BigInteger id, WebCategory entity, List<ExcludedFields> excludedFields = null);
         IList<WebCategory> GetAllCategoriesWithFields(IList<string> fields);
-        int GetMaxCategoryId();
+        BigInteger GetMaxCategoryId();
         IList<WebCategory> SearchCategories(string searchTerm, int page = 1, int pageSize = 10, int maxPage = 1);
 
         #endregion
 
         #region Customers
 
-        int CreateCustomer(WebCustomer entity);
+        BigInteger CreateCustomer(WebCustomer entity);
         IEnumerable<WebCustomer> SearchCustomers(string searchTerm, int page = 1, int pageSize = 10, int maxPage = 1);
         IEnumerable<WebCustomer> GetAllCustomersWithFields(IList<string> fields);
-        WebCustomer GetCustomerById(int id);
+        WebCustomer GetCustomerById(BigInteger id);
 
-        IEnumerable<KeyValuePair<int, string>> GetCustomerIdAndNameBySearch(string searchTerm, int page = 1, int pageSize = 10,
+        IEnumerable<KeyValuePair<BigInteger, string>> GetCustomerIdAndNameBySearch(string searchTerm, int page = 1, int pageSize = 10,
             int maxPage = 1);
         
 
@@ -61,13 +62,13 @@ namespace WebApi.Contexts.Interfaces
         
         # region Orders
         
-        IEnumerable<WebOrder> GetAllOrdersExcludeById(IEnumerable<int> idsToExclude, DateTime? startDate, DateTime? endDate);
-        IEnumerable<WebOrder> GetOrdersByFilters(DateTime? startDate, DateTime? endDate, IEnumerable<int> idsToExclude = null, IEnumerable<OrderStatus> orderStatuses = null);
+        IEnumerable<WebOrder> GetAllOrdersExcludeById(IEnumerable<BigInteger> idsToExclude, DateTime? startDate, DateTime? endDate);
+        IEnumerable<WebOrder> GetOrdersByFilters(DateTime? startDate, DateTime? endDate, IEnumerable<BigInteger> idsToExclude = null, IEnumerable<OrderStatus> orderStatuses = null);
         IEnumerable<WebOrder> GetOrdersBySearch(
-            IEnumerable<int> idsToExclude,
+            IEnumerable<BigInteger> idsToExclude,
             string searchTerm,
             IEnumerable<OrderStatus> statuses,
-            int? customerId,
+            BigInteger? customerId,
             decimal totalMin,
             decimal totalMax,
             DateTime startDate,
@@ -75,7 +76,7 @@ namespace WebApi.Contexts.Interfaces
             int page = 1,
             int pageSize = 10,
             int maxPage = 1);
-        void UpdateOrderStatus(int orderId, OrderStatus status);
+        void UpdateOrderStatus(BigInteger orderId, OrderStatus status);
         #endregion
 
         #region PaymentMethods
